requirements.txt
beautifulsoup4==4.12.3
pymongo==4.6.1
PyYAML==6.0.1
Requests==2.32.3
selenium==3.141.0
tqdm==4.66.4
